+++ src/SerializedArray.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * This file is part of SerializedArray.
+ * This file is part of SerializedArray
  *
  * Licensed under The MIT License
  * For full copyright and license information, please see the LICENSE.txt
@@ -11,6 +11,8 @@
  * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
 namespace SerializedArray;
+
+use Exception;
 
 /**
  * This class allows you to read and write arrays using text files
@@ -27,6 +29,7 @@ class SerializedArray
      * Construct
      * @param string $file Text file you want to read/write
      * @return void
+     * @throws Exception
      * @uses $file
      */
     public function __construct($file)
@@ -38,7 +41,7 @@ class SerializedArray
         }
 
         if (!is_writable($target)) {
-            trigger_error(sprintf('File or directory %s not writeable', $target), E_USER_ERROR);
+            throw new Exception(sprintf('File or directory %s not writeable', $target), E_USER_ERROR);
         }
 
         $this->file = $file;