+++ tests/bootstrap.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * This file is part of SerializedArray.
+ * This file is part of SerializedArray
  *
  * Licensed under The MIT License
  * For full copyright and license information, please see the LICENSE.txt