+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 1.0 branch
 ### 1.0.2
 * updated for phpunit 6.0;
+* `SerializedArray` throws an exception if the target is not writable;
 * the MIT license has been applied.
 
 ### 1.0.1